pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "agents>=1.4.0",
    "kubernetes>=32.0.1",
    "networkx>=3.4.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.44.1",
]
